[fix]: merge test file
